Create configuration mechanism for image manipulation
Now you can configure multiple instructions for an image processor.

For example

ImageProcessor().Resize(100,100).ApplyFilter(ImageFilter.Sepia).And.Resize(60,
null).And.Resize(0, 600)

As you can see here we instruct the processor to do three tasks (divided
by an And operator). An And operator marks the beginning of a new
instruction.