Adds additional extension to Markdown renderer to allow multiple markdig extensions (#429)
* Adds extension method to allow multiple markdown extension types to be added

Extension method really just provides ease of use for the recipes. If you
need to customize Markdig for whatever reason the current recipes don't
allow specifying anything beyond the build in extensions. This allows
whoever is consuming the recipes to specify an array of other extensions
that implement IMarkdownExtension in the settings to be applied.

Currently if you need to add a new extension you are forced into clearing
the pipeline and adding it back in with a slightly different syntax for
the markdown processing which is cumbersome.

* Adds unit tests for Markdown extension methods