Adds Sass template processing
This more or less copies the LESS module but adds Sass processing using
LibSass.Net. LibSass.Net is a wrapper around the C library LibSass which
worried me a bit, but everything seems to wire up just fine.

The test project is pretty simplistic. I would have liked to have a few
test cases for testing out that I got the include functionality right, but
I couldn't wrap my head around how to do that easily. But basic
verification of a successful call and a failing call are there for now.