Uses ThreadLocal<MsieJsEngine> to reuse JS engine
MsieJsEngine is certainly not thread safe so I was rebuildingi t from scratch within the loop. No sense in doing that, so switch the code to use a ThreadLocal instance to pool the JS instances between threads.  Not a huge win for small sites, but for a 1000 document test run it went from 15s to 1s on my machine.