Adds test for auto-highlighting
Adds a test where the code block doesn't have a language and highlightAuto
must be called. This not only covers that scenario but is an important
test case for javascript execution engines. Because highlight.js doesn't
know the language it must iterate through the languages it is configured
with and check them one by one using regular expressions. These have a
wide variety of implementations so it actually provides a good test case
for a javascript languages regex support. It seems anything doing JS -> IL
starts to fall short here due to the differences in supported Regex
escaping and language features. And with regex being a huge part of
linting and text manipulation supporting regex is a vital part of
evaluating a JS engine.