Adds extension methods to IModuleCollection to help tweak recipes
Adds a handful of helpers for inserting new modules into an existing pipeline which should be very helpful for people needing to slightly tweak recipes.

While testing this I uncovered a handful of other bugs

* ModuleCollection.Insert had no bounds check on its loop
* Null reference exception on GetNamePair when key is null

Also aAdded the key name to the KeyNotFoundException thrown from the indexer. I figure we aren't exactly exposing state secrets here so might as well throw the consumer a bone if they misstype key names as frequently as me.