Removes escaping of @ in code blocks
This was escaping ALL @ signs in the document instead of just the ones in the code blocks. Not good for things like actual javascript and the such. I tried to get AngularSharp to be nice and let me escape it via the InnerHtml but they insisted on changing it back. 

Taking a step back I realized all of this is foolish. This should really only be called after Razor is invoked. Added a test to make sure the highlighting is cool with the escaped html in a code block razor will generate.