Filters JetBrains assembly from loaded namespace list (#453)
DotTrace shoves it's profiling assembly in dynamically when running a profile of the application. Not a huge deal but it gets picked up as a loaded namespace but the tools that write the configuration file using statements and with the razor parser. If fails to find that DLL and fails. This simply filters it out if found from the config.