Java script engine switcher and highlight js (#452)
* Adds server-side code block formatting

This applies the highlight.js highlighting on the server allowing the client side script to be dropped. It does this using the MsieJavaScriptEngine to execute the script. The script itself is actually the node version, but I applied browserify to it so that it wouldn't require node anymore. Kind of a hack, but it works. I included the steps I took to build the file in regenerating-highlight-all.js so that it could be updated in the future with a newer version.

* Uses ThreadLocal<MsieJsEngine> to reuse JS engine 

MsieJsEngine is certainly not thread safe so I was rebuildingi t from scratch within the loop. No sense in doing that, so switch the code to use a ThreadLocal instance to pool the JS instances between threads.  Not a huge win for small sites, but for a 1000 document test run it went from 15s to 1s on my machine.

* Adds test for auto-highlighting

Adds a test where the code block doesn't have a language and highlightAuto
must be called. This not only covers that scenario but is an important
test case for javascript execution engines. Because highlight.js doesn't
know the language it must iterate through the languages it is configured
with and check them one by one using regular expressions. These have a
wide variety of implementations so it actually provides a good test case
for a javascript languages regex support. It seems anything doing JS -> IL
starts to fall short here due to the differences in supported Regex
escaping and language features. And with regex being a huge part of
linting and text manipulation supporting regex is a vital part of
evaluating a JS engine.

* Adds JavaScriptEngineSwitcher for javascript processing

Adds the following packages to Core

* JavaScriptEngineSwitcher - an abstraction around common JS processing libraries
* JsPool - a pool around the engine switcher. We'll be creating these a lot potentially so good pooling is a must
* Jint - managed implementation of a javascript processor. We'll make this the default.
* JavaScriptEngineSwitcher.Jint - allows Jint to be plugged into the JavaScriptEngineSwitcher

For a module that wants to use the javascript engine it's pretty simple. They just need to call GetJsEngineFromPool off of the context. When they are done then must call ReturnJsEngineToPool, This exposes more or less the same interface as IJsEngine from the JavaScriptEngineSwitcher so it should be easy to find help and examples for devs to work with. The implementation is just a wrapper of the default implementation plus a couple of helpers around ensuring libraries only get loaded once per engine.

Things get tricky in two spots

1. Testing. With this being exposed off of IExecutionContext the TestContext project needed some implementation. Rather than bring in a JS engine here too I made it a Func<> that could be set to a factory for building up the JsEngine by the test. I added a Wyam.Tests.JavaScript with an implementation of an engine that could be used for testing. It's just a copy and paste from Wyam.Core but it does allow test projects to be able to avoid needing to reference Core just to get some JS processing to work
2. Reloading the configuration. Because the configuration could potentially change the JS engine it needed the ability to wipe the singleton that JavaScriptEngineSwitcher uses. Had to add a couple static methods off of Engine and ExecutionContext to get those settings cleared out properly.

* Marks CanHighlightAutoCodeBlocks as Ignored

Test is failing due to Jint's regex implementation, but is still valuable to keep around for testing those fixes in the future

* Removes escaping of @ in code blocks

This was escaping ALL @ signs in the document instead of just the ones in the code blocks. Not good for things like actual javascript and the such. I tried to get AngularSharp to be nice and let me escape it via the InnerHtml but they insisted on changing it back. 

Taking a step back I realized all of this is foolish. This should really only be called after Razor is invoked. Added a test to make sure the highlighting is cool with the escaped html in a code block razor will generate.

* Tweaks highlight.js to have more .net compliant regex

The ADA and Lisp languages used some regex syntax that .Net couldn't handle. By changing this it allows Jint to run successfully.